Set eval = FALSE for code requiring MCMC samples
Part of the code requires the entire posterior for parameters. These can be downloaded as supplementary material for the article (https://esajournals.onlinelibrary.wiley.com/doi/full/10.1002/ecs2.3546) but are too large to stick on GitHub. 
If you download the data package and stick the file "ArcticFox_IPM_PosteriorSamples.RData" into the "Files_ArcticFox" folder, you should be able to run everything.